Move debug utilities into optional header
